requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
importlib-metadata==4.0.1
iniconfig==1.1.1
mongomock==3.22.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
pytz==2021.1
sentinels==1.0.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.4.1
